+++ spec/pg_eventstore/config_spec.rb
@@ -5,8 +5,7 @@ RSpec.describe PgEventstore::Config do
 
   let(:instance) { described_class.new(name: :my_awesome_config) }
 
-  it { is_expected.to be_a(PgEventstore::Extensions::OptionsExtension) }
-  it { expect(subject.name).to eq(:my_awesome_config) }
+  it { is_expected.to be_a(PgEventstore::BasicConfig) }
 
   describe 'attributes' do
     it do
@@ -25,17 +24,5 @@ RSpec.describe PgEventstore::Config do
     it { is_expected.to have_option(:subscription_max_retries).with_default_value(5) }
     it { is_expected.to have_option(:subscription_restart_terminator) }
     it { is_expected.to have_option(:subscription_graceful_shutdown_timeout).with_default_value(15) }
-  end
-
-  describe '#connection_options' do
-    subject { instance.connection_options }
-
-    before do
-      instance.pg_uri = 'postgresql://localhost:5432'
-      instance.connection_pool_size = 12
-      instance.connection_pool_timeout = 34
-    end
-
-    it { is_expected.to eq(uri: 'postgresql://localhost:5432', pool_size: 12, pool_timeout: 34) }
   end
 end